+++ R/read.R
@@ -27,6 +27,8 @@
 #' integer and 32-bit float samples).
 #'
 #' @param path A string. The path to the tiff file to read.
+#' @param frames Which frames do you want to read. Default all. To read the 2nd
+#'   and 7th frames, use `frames = c(2, 7)`.
 #' @param list_safety A string. This is for type safety of this function. Since
 #'   returning a list is unlikely and probably unexpected, the default is to
 #'   error. You can instead opt to throw a warning (`list_safety = "warning"`)
@@ -49,94 +51,32 @@
 #' img <- read_tif(system.file("img", "Rlogo.tif", package = "ijtiff"))
 #' img <- read_tif(system.file("img", "2ch_ij.tif", package = "ijtiff"))
 #' str(img)  # we see that `ijtiff` correctly recognises this image's 2 channels
+#' img <- read_tif(system.file("img", "2ch_ij.tif", package = "ijtiff"),
+#'                 frames = c(1, 3))
 #'
 #' @export
-read_tif <- function(path, list_safety = "error", msg = TRUE) {
-  checkmate::assert_string(path)
-  path %<>% stringr::str_replace_all(stringr::coll("\\"), "/") # windows safe
-  checkmate::assert_file_exists(path)
-  if (stringr::str_detect(path, "/")) {
-    tiff_dir <- filesstrings::str_before_last(path, "/")
-    checkmate::assert_directory_exists(tiff_dir)
-    init_wd <- setwd(tiff_dir)
-    on.exit(setwd(init_wd))
-    path %<>% filesstrings::str_after_last("/")
-    # `read_tif()` sometimes fails when writing to far away directories.
-  }
+read_tif <- function(path, frames = "all", list_safety = "error", msg = TRUE) {
+  path %<>% prep_path()
+  frames %<>% prep_frames()
+  withr::local_dir(attr(path, "path_dir"))  # setwd(attr(path, "path_dir"))
   checkmate::assert_logical(msg, max.len = 1)
   checkmate::assert_string(list_safety)
   list_safety %<>% filesstrings::match_arg(c("error", "warning", "none"),
     ignore_case = TRUE
   )
-  out <- .Call("read_tif_C", path.expand(path), PACKAGE = "ijtiff")
+  tags1 <- read_tags(path, frames = 1)[[1]]
+  prep <- prep_read(path, frames, tags1, tags = FALSE)
+  out <- .Call("read_tif_C", path.expand(path), prep$frames,
+               PACKAGE = "ijtiff") %>%
+    {.[prep$back_map]}
   checkmate::assert_list(out)
   ds <- dims(out)
   if (filesstrings::all_equal(ds)) {
     d <- ds[[1]]
     attrs1 <- attributes(out[[1]])
-    n_ch <- 1
-    if ("samples_per_pixel" %in% names(attrs1)) n_ch <- attrs1$samples_per_pixel
-    if ("description" %in% names(attrs1)) {
-      description <- attrs1$description
-      if (startsWith(description, "ImageJ")) {
-        ij_n_ch <- FALSE
-        if (stringr::str_detect(description, "channels=")) {
-          n_ch <- description %>%
-            filesstrings::str_after_first("channels=") %>%
-            filesstrings::first_number()
-          ij_n_ch <- TRUE
-        }
-        n_imgs <- NA
-        if (stringr::str_detect(description, "images=")) {
-          n_imgs <- description %>%
-            filesstrings::str_after_first("images=") %>%
-            filesstrings::first_number()
-        }
-        n_slices <- NA
-        if (stringr::str_detect(description, "slices=")) {
-          n_slices <- description %>%
-            filesstrings::str_after_first("slices=") %>%
-            filesstrings::first_number()
-        }
-        if (stringr::str_detect(description, "frames=")) {
-          n_frames <- description %>%
-            filesstrings::str_after_first("frames=") %>%
-            filesstrings::first_number()
-          if (!is.na(n_slices)) {
-            stop(
-              "The ImageJ-written image you're trying to read says it has ",
-              n_frames, " frames AND ", n_slices, " slices. To be read by ",
-              "the 'ijtiff' package, the number of slices OR the number ",
-              "of frames should be specified in the description tiff tag ",
-              "(and they're interpreted as the same thing), but not both. "
-            )
-          }
-          n_slices <- n_frames
-        }
-        if (!is.na(n_slices) && !is.na(n_imgs)) {
-          if (ij_n_ch) {
-            if (n_imgs != n_ch * n_slices) {
-              stop(
-                "The ImageJ-written image you're trying to read says in its",
-                " TIFFTAG_DESCRIPTION that it has ", n_imgs, " images of ",
-                n_slices, " slices of ", n_ch,
-                " channels. However, with ", n_slices, " slices of ", n_ch,
-                " channels, one would expect there to be ", n_slices, "x",
-                n_ch, "=", n_ch * n_slices, " images. ",
-                "This discrepancy means that the ",
-                "'ijtiff' package can't read your image correctly. One ",
-                "possible source of this kind of error is that your image ",
-                "is temporal and volumetric. 'ijtiff' can handle either ",
-                "time-based or volumetric stacks, but not both."
-              )
-            }
-          }
-        }
-        if ((isTRUE(length(out) == n_imgs) && ij_n_ch) ||
-          ((!ij_n_ch) && n_ch == 1)) {
-          if (length(d) > 2) out %<>% purrr::map(extract_desired_plane)
-        }
-      }
+    if ((isTRUE(length(out) == prep$n_imgs) && prep$ij_n_ch) ||
+        ((!prep$ij_n_ch) && prep$n_ch == 1)) {
+      if (length(d) > 2) out %<>% purrr::map(extract_desired_plane)
     }
     out %<>% unlist()
     if (attrs1$sample_format == "uint") {
@@ -160,7 +100,7 @@ read_tif <- function(path, list_safety = "error", msg = TRUE) {
         }
       }
     }
-    dim(out) <- c(d[1:2], n_ch, length(out) / prod(c(d[1:2], n_ch)))
+    dim(out) <- c(d[1:2], prep$n_ch, length(out) / prod(c(d[1:2], prep$n_ch)))
     if ("dim" %in% names(attrs1)) attrs1$dim <- NULL
     do_call_list <- c(list(img = out), attrs1)
     out <- do.call(ijtiff_img, do_call_list)
@@ -186,7 +126,7 @@ read_tif <- function(path, list_safety = "error", msg = TRUE) {
     dim(out) %>% {
       pretty_msg(
         "Reading ", path, ": ", bps, .[1], "x", .[2], " pixel image of ",
-        ifelse(ints, "unsigned integer", "floating point"), " type with ",
+        ifelse(ints, "unsigned integer", "floating point"), " type. Reading ",
         .[3], " channel", ifelse(.[3] > 1, "s", ""), " and ", .[4],
         " frame", ifelse(.[4] > 1, "s", ""), " . . ."
       )
@@ -196,16 +136,20 @@ read_tif <- function(path, list_safety = "error", msg = TRUE) {
   out
 }
 
+#' @rdname read_tif
+#' @export
+tif_read <- read_tif
+
+
 #' Read TIFF tag information without actually reading the image array.
 #'
 #' TIFF files contain metadata about images in their _TIFF tags_. This function
 #' is for reading this information without reading the actual image.
 #'
 #' @inheritParams read_tif
-#' @param all TIFF files can contain multiple images. With `all = TRUE`, the
-#'   information about all images is returned in a list of lists. To just get
-#'   the information about some images, pass those image numbers to the `all`
-#'   parameter (see examples). `all = FALSE` is equivalent to `all = 1`.
+#' @param frames Which frames do you want to read tags from. Default first frame
+#'   only. To read from the 2nd and 7th frames, use `frames = c(2, 7)`, to read
+#'   from all frames, use `frames = "all"`.
 #'
 #' @return A list of lists.
 #'
@@ -216,42 +160,31 @@ read_tif <- function(path, list_safety = "error", msg = TRUE) {
 #' @examples
 #' read_tags(system.file("img", "Rlogo.tif", package="ijtiff"))
 #' read_tags(system.file("img", "2ch_ij.tif", package="ijtiff"))
-#' read_tags(system.file("img", "2ch_ij.tif", package="ijtiff"), all = c(2, 4))
+#' read_tags(system.file("img", "2ch_ij.tif", package="ijtiff"),
+#'           frames = c(2, 4))
 #'
 #' @export
-read_tags <- function(path, all = TRUE) {
-  checkmate::assert_string(path)
-  path %<>% stringr::str_replace_all(stringr::coll("\\"), "/") # windows safe
-  checkmate::assert_file_exists(path)
-  if (stringr::str_detect(path, "/")) {
-    init_wd <- setwd(filesstrings::str_before_last(path, "/"))
-    on.exit(setwd(init_wd))
-    path %<>% filesstrings::str_after_last("/")
-    # `read_tags()` sometimes fails when writing to far away directories.
+read_tags <- function(path, frames = 1) {
+  frames %<>% prep_frames()
+  path %<>% prep_path()
+  withr::local_dir(attr(path, "path_dir"))  # setwd(attr(path, "path_dir"))
+  if (isTRUE(all.equal(frames, 1,
+                       check.attributes = FALSE, check.names = FALSE))) {
+    return(
+      list(frame1 = .Call("read_tags_C", path, 1L, PACKAGE = "ijtiff")[[1]])
+    )
   }
-  checkmate::assert(
-    checkmate::check_flag(all),
-    checkmate::check_integerish(all, lower = 1)
-  )
-  if (is.numeric(all)) {
-    n_tiff_dirs <- count_imgs(path)
-    all_max <- max(all)
-    if (all_max > n_tiff_dirs) {
-      custom_stop(
-        "Cannot access image {all_max}.",
-        "
-         You have tried to access information from image {all_max},
-         but there are only {n_tiff_dirs} images in total.
-        "
-      )
-    }
-  }
-  if (rlang::is_false(all)) all <- 1
-  out <- .Call("read_tags_C", path, all, PACKAGE = "ijtiff")
-  if (rlang::is_true(all)) {
-    names(out) <- paste0("img", seq_along(out))
-  } else {
-    names(out) <- paste0("img", all)
-  }
+  tags1 <- read_tags(path, frames = 1)[[1]]
+  prep <- prep_read(path, frames, tags1, tags = TRUE)
+  out <- .Call("read_tags_C", path, prep$frames, PACKAGE = "ijtiff") %>%
+    {.[prep$back_map]}
+  frame_nums <- prep$frames[prep$back_map]
+  if (!is.na(prep$n_slices) && prep$n_dirs != prep$n_slices)
+    frame_nums <- ceiling(frame_nums / prep$n_ch)
+  names(out) <- paste0("frame", filesstrings::nice_nums(frame_nums))
   out
 }
+
+#' @rdname read_tags
+#' @export
+tags_read <- read_tags